fix: fixed gameplay bugs (#197)
* fixed warnings in update function

* Made some changes to how the door panel works
Simplified the AI's door handling behavior until this behavior is fixed
Fixed bug that would cause random crashes when loading animation data in the constant buffer (needs further testing for proof of fix)

* Fixed bug where neighbor of door was not getting set up properly after calling dissable.

* Fixed bug where AI would kill player upon entering the room without being near it

* fixed bug where dead animation would play on retry after continue
bug cause: animation data was not being cleaned properly thus making it so that the same data from the previous run was used instead

* path finder working as intended again

* camera data now only exist in 1 place rather than in every single room

* all rooms no longer generate their own floor

* rooms no longer create an entity per wall and door.
Instead 1 room and 4 walls are made which are shared among all rooms via index
Removed recent debugging code used in the path finder / tiles

* Doors are finalized and cannot be soft locked any more

* added health to interactibles so that the AI has to hit them a few times before breaking them
moved right click functionality to left click

* fixed bug where player would not walk the first time it clicked on an interactible
details: walking speed was being initiated to 0 rather than the default walking speed

fixed bug in left click after rework where player would move after clicked area under minimap

* Bug fix: enemy AI stops attacking player after player has died
Just made it so that the AI stopped targeting the player if the player was already dead

* adjusted clean up on unload of entities to preserve wall and floor meshes

+++ src/cpp/player_behavior.cpp
@@ -88,8 +88,8 @@ void set_velocity(float2e destination, crow::entities& entity, size_t index,
 void set_velocity(float2e position, float2e destination,
                   float3e& velocity, float speed) {
   float rad_angle = crow::get_angle(position, destination);
-  float vel_y = std::sin(rad_angle) * speed;
-  float vel_x = std::cos(rad_angle) * speed;
+  float vel_y = std::sinf(rad_angle) * speed;
+  float vel_x = std::cosf(rad_angle) * speed;
   velocity = float3e{vel_x, 0, vel_y};
 }
